Add files via upload
Sensor Based Waste Collection Bins is used to identify status of waste bins if it is empty or filled so as to customize the waste
collection schedule accordingly and also save the cost. Real time waste management system by using smart dustbins to check the
fill level of dustbins whether the dustbins are full or not, through this system the information of all smart dustbins can be accessed
from anywhere and anytime by the concern person. It will inform the status of each and every dustbin in real time so that
concerned authority can send the garbage collection vehicle only when the dustbin is full. By implementing this system resource
optimization, cost reduction, effective usage of smart dustbins can be done. Scribes the scope of work of smart bins in managing
the waste collection system for an entire city. The network of sensors enabled smart bins connected through the cellular network
generates a large amount of data, which is further analyzed and visualized at real time to gain insights about the status of waste
around the city. The scope for the future work of this system can be implement on various places like Bus stop, Railway stations
etc. As shown in below figure it will display the Backside street information sticker, there will be a facilities like , LED Signboard
and the Bio Enable Smarting sensor inside the dustbin.